+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/PlayerProfile.java
@@ -68,7 +68,7 @@ public class PlayerProfile {
         this.lastIp = lastIp;
     }
 
-    public long getLastLogin() {
+    public synchronized long getLastLogin() {
         return lastLogin;
     }