+++ bungee/src/main/java/com/github/games647/fastlogin/bungee/MojangApiConnector.java
@@ -1,6 +1,7 @@
 package com.github.games647.fastlogin.bungee;
 
 import com.google.gson.Gson;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -9,6 +10,7 @@ import java.net.URL;
 import java.util.UUID;
 import java.util.logging.Level;
 import java.util.regex.Pattern;
+
 import net.md_5.bungee.BungeeCord;
 
 public class MojangApiConnector {